remove confusing styles.tsx
