+++ packages/itwinui-react/src/core/utils/components/Resizer.test.tsx
@@ -5,7 +5,7 @@
 import * as React from 'react';
 import { fireEvent, render } from '@testing-library/react';
 import Resizer, { type ResizerProps } from './Resizer.js';
-import * as Styles from '../functions/styles.js';
+import * as Styles from '../functions/dom.js';
 
 jest.spyOn(Styles, 'getTranslateValues').mockReturnValue([100, 100]);
 jest.spyOn(window, 'getComputedStyle').mockReturnValue({
