+++ packages/itwinui-react/src/core/utils/functions/dom.ts
@@ -56,3 +56,20 @@ export const mergeEventHandlers =
       }
     }
   };
+
+/**
+ * Returns translate values as an array of a given element.
+ * @param element HTML element you want to get translate value of
+ * @returns Translate values in pixels in an array `[translateX, translateY]`
+ */
+export const getTranslateValues = (element: HTMLElement | null | undefined) => {
+  if (!element) {
+    return [];
+  }
+
+  const transformValue =
+    getComputedStyle(element).getPropertyValue('transform');
+  const matrix = new DOMMatrix(transformValue);
+
+  return [matrix.m41, matrix.m42];
+};
